refactor 1단계 미션 코드 리팩토링
 - package 명 study -> basic 으로 변경
 - 테스트 대상이 되는 코드는 main 디렉토리 이하로 이동
