+++ src/test/kotlin/basic/PersonTest.kt
@@ -1,8 +1,9 @@
-package study
+package basic
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 
+@Suppress("SpellCheckingInspection")
 class PersonTest {
 
     @Test
